move in-app stuff into KumulosInApp
