+++ kumulos/src/main/java/com/kumulos/android/InAppJavaScriptInterface.java
@@ -83,8 +83,8 @@ class InAppJavaScriptInterface {
                     this.openUrl(currentActivity, action.getUrl());
                     return;
                 case BUTTON_ACTION_DEEP_LINK:
-                    if (Kumulos.inAppDeepLinkHandler != null){
-                        Kumulos.inAppDeepLinkHandler.handle(action.getDeepLink());
+                    if (KumulosInApp.inAppDeepLinkHandler != null){
+                        KumulosInApp.inAppDeepLinkHandler.handle(action.getDeepLink());
                     }
                     return;
                 case BUTTON_ACTION_REQUEST_APP_STORE_RATING:
